Implement isDeviceSupportedAsResizeTarget
