+++ API.md
@@ -12,14 +12,14 @@
 <dt><a href="#open">async open(browserInfo, pageUrl)</a></dt>
 <dd><p>Opens the web page in a new instance of the browser.</p>
 </dd>
+<dt><a href="#screenshot">async screenshot(pageUrl, screenshotPath)</a></dt>
+<dd><p>Takes a screenshot of the browser window where the specified web page is opened.</p>
+</dd>
 <dt><a href="#resize">async resize(pageUrl, width, height)</a></dt>
 <dd><p>Changes the browser window size to the new width and height.</p>
 </dd>
 <dt><a href="#resize">async resize(pageUrl, deviceName, [orientation])</a></dt>
 <dd><p>Changes the browser window size according to the screen size of the target device.</p>
-</dd>
-<dt><a href="#screenshot">async screenshot(pageUrl, screenshotPath)</a></dt>
-<dd><p>Takes a screenshot of the browser window where the specified web page is opened.</p>
 </dd>
 </dl>
 ## Typedefs
@@ -54,23 +54,10 @@ Returns information about the specified browser.
 Returns the list of the [BrowserInfo](#BrowserInfo) objects that contain information about the browsers installed on the machine.
 
 **Kind**: global [async](http://tc39.github.io/ecmascript-asyncawait/) function  
-**Returns**: <code>Object.&lt;string, BrowserInfo&gt;</code> - List of the [BrowserInfo](#BrowserInfo) objects
-  containing information about the browsers installed on the machine.  
+**Returns**: <code>Object.&lt;string, BrowserInfo&gt;</code> - List of the [BrowserInfo](#BrowserInfo) objects  containing information about the browsers installed on the machine.  
 **Example**  
 ```js
-{
-  chrome: {
-      path: 'C:\\ProgramFiles\\...\\chrome.exe',
-      cmd: '--new-window',
-      macOpenCmdTemplate: 'open -n -a "{{{path}}}" --args {{{pageUrl}}} {{{cmd}}}'
-  },
-
-  firefox: {
-      path: 'C:\\ProgramFiles\\...\\firefox.exe',
-      cmd: '-new-window',
-      macOpenCmdTemplate: 'open -a "{{{path}}}" {{{pageUrl}}} --args {{{cmd}}}'
-  }
-}
+{  chrome: {      path: 'C:\\ProgramFiles\\...\\chrome.exe',      cmd: '--new-window',      macOpenCmdTemplate: 'open -n -a "{{{path}}}" --args {{{pageUrl}}} {{{cmd}}}'  },  firefox: {      path: 'C:\\ProgramFiles\\...\\firefox.exe',      cmd: '-new-window',      macOpenCmdTemplate: 'open -a "{{{path}}}" {{{pageUrl}}} --args {{{cmd}}}'  }}
 ```
 <a name="open"></a>
 ## *async* open(browserInfo, pageUrl)
@@ -82,6 +69,17 @@ Opens the web page in a new instance of the browser.
 | --- | --- | --- |
 | browserInfo | <code>[BrowserInfo](#BrowserInfo)</code> | Provides information on the browser where the web page should be opened. |
 | pageUrl | <code>string</code> | Specifies the web page URL. |
+
+<a name="screenshot"></a>
+## *async* screenshot(pageUrl, screenshotPath)
+Takes a screenshot of the browser window where the specified web page is opened.
+
+**Kind**: global [async](http://tc39.github.io/ecmascript-asyncawait/) function  
+
+| Param | Type | Description |
+| --- | --- | --- |
+| pageUrl | <code>string</code> | Specifies the URL of the web page opened in the browser. |
+| screenshotPath | <code>string</code> | Specifies the full path to the screenshot file. For example, D:\Temp\chrome-screenshot.jpg. |
 
 <a name="resize"></a>
 ## *async* resize(pageUrl, width, height)
@@ -107,17 +105,6 @@ Changes the browser window size according to the screen size of the target devic
 | deviceName | <code>string</code> |  | Specifies the name of the target device. You can use the values specified in the Device Name column of [this table](http://viewportsizes.com/). |
 | [orientation] | <code>string</code> | <code>&quot;landscape&quot;</code> | Specifies the device orientation: "portrait" or "landscape". |
 
-<a name="screenshot"></a>
-## *async* screenshot(pageUrl, screenshotPath)
-Takes a screenshot of the browser window where the specified web page is opened.
-
-**Kind**: global [async](http://tc39.github.io/ecmascript-asyncawait/) function  
-
-| Param | Type | Description |
-| --- | --- | --- |
-| pageUrl | <code>string</code> | Specifies the URL of the web page opened in the browser. |
-| screenshotPath | <code>string</code> | Specifies the full path to the screenshot file. For example, D:\Temp\chrome-screenshot.jpg. |
-
 <a name="BrowserInfo"></a>
 ## BrowserInfo : <code>Object</code>
 Object that contains information about the browser installed on the machine.
@@ -134,9 +121,5 @@ Object that contains information about the browser installed on the machine.
 
 **Example**  
 ```js
-{
-      path: 'C:\\ProgramFiles\\...\\firefox.exe',
-      cmd: '-new-window',
-      macOpenCmdTemplate: 'open -a "{{{path}}}" {{{pageUrl}}} --args {{{cmd}}}'
- }
+{      path: 'C:\\ProgramFiles\\...\\firefox.exe',      cmd: '-new-window',      macOpenCmdTemplate: 'open -a "{{{path}}}" {{{pageUrl}}} --args {{{cmd}}}' }
 ```
