+++ src/api/resize/client-area.js
@@ -1,0 +1,3 @@
+/*
+ TODO: Implement client area resizing
+ */
