+++ src/index.js
@@ -1,7 +1,8 @@
 import getInstallations from './api/get-installations';
 import open from './api/open';
 import close from './api/close';
-import resize from './api/resize';
+import resize from './api/resize/window';
+import { isDeviceSupported as isDeviceSupportedAsResizeTarget } from './api/resize/utils';
 import screenshot from './api/screenshot';
 import getBrowserInfo from './api/get-browser-info';
 
@@ -12,5 +13,6 @@ export default {
     open,
     close,
     resize,
-    screenshot
+    screenshot,
+    isDeviceSupportedAsResizeTarget
 };
